Split district_policies.rs into subdirectory modules
Split the 1258-line district_policies.rs into 6 files under
district_policies/, each under the 500-line limit:
- mod.rs (29 lines): module declarations and re-exports
- types.rs (220 lines): constants, DistrictPolicyOverrides, DistrictPolicyState
- lookup.rs (246 lines): DistrictPolicyLookup, pure helper functions
- systems.rs (140 lines): ECS system, Saveable impl, plugin
- tests.rs (394 lines): type/state/helper unit tests
- tests_lookup.rs (268 lines): lookup and saveable unit tests

All public items re-exported from mod.rs so no caller changes needed.

Part of #1389
